V17/image upload mapper (#959)
* update packages

* Split image upload and Image cropper mappers (to handle uploads in media slightly diffrently)

* Potential fix for pull request finding


* Potential fix for pull request finding


* add missing xml comments

* add debug , so we can see which mapper we hit.

---------
