+++ Directory.Packages.props
@@ -1,36 +1,36 @@
 <Project>
-  <PropertyGroup>
-    <ManagePackageVersionsCentrally>true</ManagePackageVersionsCentrally>
-  </PropertyGroup>
-  <!-- umbraco site -->
-  <ItemGroup>
-    <PackageVersion Include="Our.Umbraco.ConditionalDisplayers" Version="16.6.0" />
-    <PackageVersion Include="Umbraco.Cms" Version="17.3.0" />
-    <PackageVersion Include="Microsoft.ICU.ICU4C.Runtime" Version="72.1.0.3" />
-    <PackageVersion Include="Umbraco.Cms.DevelopmentMode.Backoffice" Version="17.3.0" />
-    <PackageVersion Include="Clean.Core" Version="7.0.5" />
-  </ItemGroup>
-  <!-- umbraco libraries for packages -->
-  <ItemGroup>
-    <PackageVersion Include="Umbraco.Cms.Core" Version="17.3.0" />
-    <PackageVersion Include="Umbraco.Cms.Web.Website" Version="17.3.0" />
-    <PackageVersion Include="Umbraco.Cms.Api.Management" Version="17.3.0" />
-  </ItemGroup>
-  <!-- source link -->
-  <ItemGroup>
-    <PackageVersion Include="Microsoft.SourceLink.GitHub" Version="10.0.201" />
-  </ItemGroup>
-  <!-- testing -->
-  <ItemGroup>
-    <PackageVersion Include="Umbraco.Cms.Tests" Version="17.3.0" />
-    <PackageVersion Include="Umbraco.Cms.Tests.Integration" Version="17.3.0" />
-    <PackageVersion Include="Microsoft.NET.Test.Sdk" Version="18.4.0" />
-    <PackageVersion Include="NUnit3TestAdapter" Version="6.2.0" />
-    <PackageVersion Include="coverlet.collector" Version="8.0.1" />
-  </ItemGroup>
-  <!-- command line -->
-  <ItemGroup>
-    <PackageVersion Include="NJsonSchema" Version="11.5.2" />
-    <PackageVersion Include="CommandLineParser" Version="2.9.1" />
-  </ItemGroup>
+	<PropertyGroup>
+		<ManagePackageVersionsCentrally>true</ManagePackageVersionsCentrally>
+	</PropertyGroup>
+	<!-- umbraco site -->
+	<ItemGroup>
+		<PackageVersion Include="Umbraco.Cms" Version="17.4.2" />
+		<PackageVersion Include="Umbraco.Cms.DevelopmentMode.Backoffice" Version="17.4.2" />
+		<PackageVersion Include="Our.Umbraco.ConditionalDisplayers" Version="16.6.0" />
+		<PackageVersion Include="Microsoft.ICU.ICU4C.Runtime" Version="72.1.0.3" />
+		<PackageVersion Include="Clean.Core" Version="7.0.5" />
+	</ItemGroup>
+	<!-- umbraco libraries for packages -->
+	<ItemGroup>
+		<PackageVersion Include="Umbraco.Cms.Core" Version="17.3.0" />
+		<PackageVersion Include="Umbraco.Cms.Web.Website" Version="17.3.0" />
+		<PackageVersion Include="Umbraco.Cms.Api.Management" Version="17.3.0" />
+	</ItemGroup>
+	<!-- source link -->
+	<ItemGroup>
+		<PackageVersion Include="Microsoft.SourceLink.GitHub" Version="10.0.201" />
+	</ItemGroup>
+	<!-- testing -->
+	<ItemGroup>
+		<PackageVersion Include="Umbraco.Cms.Tests" Version="17.3.0" />
+		<PackageVersion Include="Umbraco.Cms.Tests.Integration" Version="17.3.0" />
+		<PackageVersion Include="Microsoft.NET.Test.Sdk" Version="18.4.0" />
+		<PackageVersion Include="NUnit3TestAdapter" Version="6.2.0" />
+		<PackageVersion Include="coverlet.collector" Version="8.0.1" />
+	</ItemGroup>
+	<!-- command line -->
+	<ItemGroup>
+		<PackageVersion Include="NJsonSchema" Version="11.5.2" />
+		<PackageVersion Include="CommandLineParser" Version="2.9.1" />
+	</ItemGroup>
 </Project>
