+++ uSync.History/history-client/package.json
@@ -1,6 +1,6 @@
 {
   "name": "usync-history-client",
-  "version": "17.3.2",
+  "version": "17.3.3",
   "licence": "Custom",
   "description": "uSync history function",
   "type": "module",
